// src/lib/moetruyen/model/getV2TeamsById200Data.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Moetruyen Public API
 * Unofficial read-only REST API for [MoeTruyen](https://moetruyen.net/).

## Quick start

Use the `/v2` routes for the current stable surface.

## Request notes

- All API endpoints have a global rate limit of 7 requests per second per IP.
- Include a valid `Origin` header such as `https://suicaodex.com` or `https://moetruyen.net` when making browser-like requests.

## Versioning

- From 01/04/2026, v1 routes are deprecated, pleade use v2 instead.

## Repositories

- API repo: [TNTKien/moetruyen-public-api](https://github.com/TNTKien/moetruyen-public-api)
- Original site repo: [dex593/web1 (MoeTruyen)](https://github.com/dex593/web1)
 * OpenAPI spec version: 0.2.0
 */

export type GetV2TeamsById200Data = {
  /**
   * @maximum 9007199254740991
   * @exclusiveMinimum 0
   */
  id: number;
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  name: string;
  intro: string | null;
  avatarUrl: string | null;
  coverUrl: string | null;
  /**
   * @minimum 0
   * @maximum 9007199254740991
   */
  memberCount: number;
  /**
   * @minimum 0
   * @maximum 9007199254740991
   */
  leaderCount: number;
  /**
   * @minimum 0
   * @maximum 9007199254740991
   */
  totalMangaCount: number;
  /**
   * @minimum 0
   * @maximum 9007199254740991
   */
  totalChapterCount: number;
  /**
   * @minimum 0
   * @maximum 9007199254740991
   */
  totalCommentCount: number;
};