// src/lib/moetruyen/model/getV2TeamsByIdMangaSort.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Moetruyen Public API
 * Unofficial read-only REST API for [MoeTruyen](https://moetruyen.net/).

## Quick start

Use the `/v2` routes for the current stable surface.

## Request notes

- All API endpoints have a global rate limit of 7 requests per second per IP.
- Include a valid `Origin` header such as `https://suicaodex.com` or `https://moetruyen.net` when making browser-like requests.

## Versioning

- From 01/04/2026, v1 routes are deprecated, pleade use v2 instead.

## Repositories

- API repo: [TNTKien/moetruyen-public-api](https://github.com/TNTKien/moetruyen-public-api)
- Original site repo: [dex593/web1 (MoeTruyen)](https://github.com/dex593/web1)
 * OpenAPI spec version: 0.2.0
 */

export type GetV2TeamsByIdMangaSort =
  (typeof GetV2TeamsByIdMangaSort)[keyof typeof GetV2TeamsByIdMangaSort];

export const GetV2TeamsByIdMangaSort = {
  updated_at: "updated_at",
  title: "title",
  popular: "popular",
} as const;